requirements.txt
requests == 2.25.1
pandas == 1.3.1
boto3 == 1.20.43
python-decouple == 3.5
black == 19.10b0
pylint == 2.9.6
pytest == 6.2.4
uvicorn == 0.17.0.post1
fastapi == 0.73.0
transformers == 4.16.0
numpy == 1.20.3
torch == 1.9.0
Jinja2 == 3.0.3
searchtweets-v2 == 1.1.1